moving some event repetition handling in Room

+++ app/src/main/kotlin/com/simplemobiletools/calendar/pro/models/EventRepetition.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "event_repetitions", indices = [(Index(value = ["id"], unique = true))])
 data class EventRepetition(
         @PrimaryKey(autoGenerate = true) var id: Long?,
-        @ColumnInfo(name = "event_id") val eventId: Int,
+        @ColumnInfo(name = "event_id") val eventId: Long,
         @ColumnInfo(name = "repeat_interval") val repeatInterval: Int,
         @ColumnInfo(name = "repeat_rule") val repeatRule: Int,
         @ColumnInfo(name = "repeat_limit") val repeatLimit: Int)
